refactor(aztec-node): split server.ts into factory + focused modules (#24283)
## Context

`aztec-node/src/aztec-node/server.ts` had grown to ~2100 lines, mixing
node construction, world-state/witness queries, block and checkpoint
reads, and tx-receipt assembly into one file. This follows the recent
`NodePublicCallsSimulator` extraction to keep the file navigable.

## Approach

`AztecNodeService` keeps its public and RPC surface unchanged — every
extracted method now delegates to a focused collaborator. The work is
split into five self-contained, individually-buildable commits:

- Move `createAndSync` to a free `createAztecNodeService` factory in
`factory.ts` (mirrors `createProverNode`), and convert the
~28-positional constructor to a single `AztecNodeServiceDeps` object.
- Extract the world-state/witness query cluster into
`modules/node_world_state_queries.ts`.
- Extract the
`normalizeBlockParameter`/`isBlockTag`/`isCheckpointTag`/`resolveCheckpointParameter`
block-parameter helpers as pure free functions in
`modules/block_parameter.ts`.
- Extract block and checkpoint reads into
`modules/node_block_provider.ts`.
- Extract `getTxReceipt` and receipt assembly into
`modules/node_tx_receipt.ts`.

The deps-object constructor also surfaced a latent bug in the TXE state
machine, which was passing `VERSION`/`CHAIN_ID` into the
`l1ChainId`/`version` slots swapped (harmless until now since both were
`1`); fixed here.
